+++ Robust/src/Benchmarks/SingleTM/SSCA2/GraphSDG.java
@@ -1,11 +1,11 @@
 public class GraphSDG {
-  public long[] startVertex;
-  public long[] endVertex;
-  public long[] intWeight;
+  public int[] startVertex;
+  public int[] endVertex;
+  public int[] intWeight;
 
   /* The idea is to store the index of the string weights (as a negative value)
-   * in the long Weight array. A negative value because we need to sort on
-   * the intWeights in Kernel 2. Hence the long long
+   * in the int Weight array. A negative value because we need to sort on
+   * the intWeights in Kernel 2. Hence the int int
    */
   public char[] strWeight;
   public int numEdgesPlaced;
